+++ tasks/copyCss.mjs
@@ -64,12 +64,19 @@ const copyCss = async ({ mode, watchEvent, watchPath }) => {
 
   try {
     // もし監視イベントが削除だった場合
-    if (watchEvent === 'unlink' || watchEvent === 'unlinkDir') {
+    if (
+      (watchEvent === 'unlink' || watchEvent === 'unlinkDir') &&
+      !path.basename(watchPath).startsWith('_')
+    ) {
       deleteDist();
     } else {
       // 監視タスクからの場合は、監視で検知したファイルのみをコピー（監視タスクで受け取るwatchPathがあるかないかで判断してる）
       if (watchPath) {
-        if ((watchEvent === 'add' || watchEvent === 'change') && watchPath.endsWith(extension)) {
+        if (
+          (watchEvent === 'add' || watchEvent === 'change') &&
+          watchPath.endsWith(extension) &&
+          !path.basename(watchPath).startsWith('_')
+        ) {
           // 監視イベントが追加か変更かつ、それがCSSファイルだった場合に実行
           await copyCssFile({ copyFilePath: watchPath });
           // await console.log(chalk.green('CSS Copy processing task completed.'));
