+++ tasks/copyHtml.mjs
@@ -30,7 +30,7 @@ const copyHtml = async ({ mode, watchEvent, watchPath }) => {
       // await fs.copyFileSync(inputPath, outputPath);
 
       // ここからPrettier関連
-      const inputFileData  = await fs.readFileSync(inputPath, 'utf8');
+      const inputFileData = await fs.readFileSync(inputPath, 'utf8');
 
       // Prettierを使用してHTMLコードを整形
       const formattedResult = await prettier.format(inputFileData, { parser: 'html' });
@@ -62,12 +62,19 @@ const copyHtml = async ({ mode, watchEvent, watchPath }) => {
 
   try {
     // もし監視イベントが削除だった場合
-    if (watchEvent === 'unlink' || watchEvent === 'unlinkDir') {
+    if (
+      (watchEvent === 'unlink' || watchEvent === 'unlinkDir') &&
+      !path.basename(watchPath).startsWith('_')
+    ) {
       deleteDist();
     } else {
       // 監視タスクからの場合は、監視で検知したファイルのみをコピー（監視タスクで受け取るwatchPathがあるかないかで判断してる）
       if (watchPath) {
-        if ((watchEvent === 'add' || watchEvent === 'change') && watchPath.endsWith(extension)) {
+        if (
+          (watchEvent === 'add' || watchEvent === 'change') &&
+          watchPath.endsWith(extension) &&
+          !path.basename(watchPath).startsWith('_')
+        ) {
           // 監視イベントが追加か変更かつ、それがHTMLファイルだった場合に実行
           await copyHtmlFile({ copyFilePath: watchPath });
           // await console.log(chalk.green('HTML Copy processing task completed.'));
